# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AI Banking Regulatory Compliance & Audit Intelligence Platform API
servers:
  - url: /api
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      responses:
        "200":
          description: Indexed documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

    post:
      operationId: uploadDocument
      tags: [documents]
      parameters:
        - $ref: "#/components/parameters/Category"
        - name: is_synthetic
          in: query
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUpload"
      responses:
        "201":
          description: Indexed document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "204":
          description: Document and its embeddings deleted
  /assistant/analyze:
    post:
      operationId: analyzeCompliance
      tags: [assistant]
      summary: Compare a selected regulation with a selected policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantRequest"
      responses:
        "200":
          description: Grounded compliance analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
  /runs:
    get:
      operationId: listAgentRuns
      tags: [agents]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Agent execution history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRun"
  /issues:
    get:
      operationId: listComplianceIssues
      tags: [issues]
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Compliance issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceIssue"

    post:
      operationId: createComplianceIssue
      tags: [issues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceIssueInput"
      responses:
        "201":
          description: Compliance issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceIssue"
  /issues/{issueId}:
    patch:
      operationId: updateComplianceIssue
      tags: [issues]
      parameters:
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceIssueUpdate"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceIssue"
  /reports:
    get:
      operationId: listAuditReports
      tags: [reports]
      responses:
        "200":
          description: Audit reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditReport"

    post:
      operationId: generateAuditReport
      tags: [reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditReportInput"
      responses:
        "201":
          description: Audit report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditReport"
  /notifications:
    post:
      operationId: sendComplianceNotification
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  parameters:
    Category:
      name: category
      in: query
      required: true
      schema:
        type: string
        enum:
          - regulation
          - policy
          - audit
          - other
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IssueId:
      name: issueId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    DocumentUpload:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
    Document:
      type: object
      required:
        - id
        - name
        - document_type
        - category
        - status
        - chunk_count
        - is_synthetic
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        document_type:
          type: string
        category:
          type: string
        status:
          type: string
        chunk_count:
          type: integer
        is_synthetic:
          type: boolean
        created_at:
          type: string
          format: date-time
        indexed_at:
          type:
            - string
            - "null"
          format: date-time
    Evidence:
      type: object
      required:
        - document_name
        - excerpt
        - source_type
      properties:
        chunk_id:
          type:
            - string
            - "null"
          format: uuid
        document_id:
          type:
            - string
            - "null"
          format: uuid
        document_name:
          type: string
        page_number:
          type:
            - integer
            - "null"
        excerpt:
          type: string
        source_type:
          type: string
        relevance_score:
          type:
            - number
            - "null"
    AssistantRequest:
      type: object
      required:
        - question
        - regulation_document_id
        - policy_document_id
      properties:
        question:
          type: string
          minLength: 10
        regulation_document_id:
          type: string
          format: uuid
        policy_document_id:
          type: string
          format: uuid
    AssistantResponse:
      type: object
      required:
        - run
        - answer
        - compliance_status
        - risk_level
        - evidence
        - source_documents
        - agents_involved
      properties:
        run:
          $ref: "#/components/schemas/AgentRun"
        answer:
          type: string
        compliance_status:
          type: string
          enum:
            - COMPLIANT
            - PARTIALLY_COMPLIANT
            - NON_COMPLIANT
            - UNVERIFIED
        risk_level:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        source_documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        agents_involved:
          type: array
          items:
            type: string
    ComplianceIssueInput:
      type: object
      required:
        - finding
        - regulation
        - risk
        - recommendation
      properties:
        finding:
          type: string
        regulation:
          type: string
        risk:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
        recommendation:
          type: string
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        source_document_ids:
          type: array
          items:
            type: string
            format: uuid
    ComplianceIssueUpdate:
      type: object
      properties:
        status:
          type:
            - string
            - "null"
          enum:
            - OPEN
            - IN_REVIEW
            - REMEDIATED
            - ACCEPTED
            - null
        risk:
          type:
            - string
            - "null"
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
            - null
        recommendation:
          type:
            - string
            - "null"
    ComplianceIssue:
      allOf:
        - $ref: "#/components/schemas/ComplianceIssueInput"
        - type: object
          required:
            - id
            - status
            - created_at
            - updated_at
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    AgentRun:
      type: object
      required:
        - id
        - thread_id
        - question
        - status
        - agents_involved
        - started_at
      properties:
        id:
          type: string
          format: uuid
        thread_id:
          type: string
        question:
          type: string
        status:
          type: string
        current_agent:
          type:
            - string
            - "null"
        agents_involved:
          type: array
          items:
            type: string
        trace_url:
          type:
            - string
            - "null"
        result:
          type:
            - object
            - "null"
        error:
          type:
            - string
            - "null"
        started_at:
          type: string
          format: date-time
        completed_at:
          type:
            - string
            - "null"
          format: date-time
    AuditReport:
      type: object
      required:
        - id
        - title
        - status
        - summary
        - findings
        - generated_by
        - created_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        summary:
          type: string
        findings:
          type: array
          items:
            type: object
            additionalProperties: true
        generated_by:
          type: string
        created_at:
          type: string
          format: date-time
    AuditReportInput:
      type: object
      required:
        - title
        - summary
      properties:
        title:
          type: string
        summary:
          type: string
        findings:
          type: array
          items:
            type: object
            additionalProperties: true
        generated_by:
          type: string
          default: Audit Report Agent
    NotificationInput:
      type: object
      required:
        - issue_id
        - message
        - recipient
      properties:
        issue_id:
          type: string
          format: uuid
        message:
          type: string
        recipient:
          type: string
    Notification:
      type: object
      required:
        - id
        - issue_id
        - recipient
        - message
        - status
        - sent_at
      properties:
        id:
          type: string
          format: uuid
        issue_id:
          type: string
          format: uuid
        recipient:
          type: string
        message:
          type: string
        status:
          type: string
        sent_at:
          type: string
          format: date-time
...